Dashboard updates
Combined common code between Task and Process dashboard classes, avoiding code duplication
UI styles update, manly to allow content to be responsive
