+++ jbpm-console-ng-dashboard/jbpm-console-ng-dashboard-client/src/main/java/org/jbpm/dashboard/renderer/client/panel/TaskDashboard.java
@@ -1,6 +1,5 @@
-/**
+/*
  * Copyright (C) 2015 Red Hat, Inc. and/or its affiliates.
-
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,7 +23,6 @@ import javax.enterprise.event.Observes;
 import javax.inject.Inject;
 
 import com.google.gwt.user.client.ui.IsWidget;
-import com.google.gwt.user.client.ui.Widget;
 import org.dashbuilder.dataset.DataSet;
 import org.dashbuilder.dataset.client.DataSetClientServices;
 import org.dashbuilder.dataset.filter.DataSetFilter;
@@ -33,7 +31,6 @@ import org.dashbuilder.dataset.group.Interval;
 import org.dashbuilder.displayer.DisplayerSettings;
 import org.dashbuilder.displayer.client.AbstractDisplayer;
 import org.dashbuilder.displayer.client.AbstractDisplayerListener;
-import org.dashbuilder.displayer.client.DataSetHandlerImpl;
 import org.dashbuilder.displayer.client.Displayer;
 import org.dashbuilder.displayer.client.DisplayerCoordinator;
 import org.dashbuilder.displayer.client.DisplayerListener;
@@ -44,19 +41,17 @@ import org.jbpm.console.ng.ht.model.events.TaskSelectionEvent;
 import org.jbpm.dashboard.renderer.client.panel.events.ProcessDashboardFocusEvent;
 import org.jbpm.dashboard.renderer.client.panel.events.TaskDashboardFocusEvent;
 import org.jbpm.dashboard.renderer.client.panel.formatter.DurationFormatter;
-import org.jbpm.dashboard.renderer.client.panel.i18n.DashboardI18n;
 import org.jbpm.dashboard.renderer.client.panel.widgets.ProcessBreadCrumb;
 import org.uberfire.client.mvp.PlaceManager;
 import org.uberfire.client.mvp.PlaceStatus;
 import org.uberfire.mvp.Command;
 
 import static org.jbpm.dashboard.renderer.model.DashboardData.*;
-import static org.uberfire.commons.validation.PortablePreconditions.*;
 
 @Dependent
-public class TaskDashboard implements IsWidget {
+public class TaskDashboard extends AbstractDashboard implements IsWidget {
 
-    public interface View extends IsWidget {
+    public interface View extends AbstractDashboard.View {
 
         void init(TaskDashboard presenter,
                   Displayer totalMetric,
@@ -78,33 +73,11 @@ public class TaskDashboard implements IsWidget {
                   Displayer tasksByStatus,
                   Displayer tasksTable);
 
-        void showBreadCrumb(String processName);
-
-        void hideBreadCrumb();
-
-        void setHeaderText(String text);
-
-        void showLoading();
-
-        void hideLoading();
-
-        void showDashboard();
-
-        void showInstances();
-
-        DashboardI18n getI18nService();
     }
 
     protected View view;
-    protected ProcessBreadCrumb processBreadCrumb;
-    protected DashboardFactory dashboardFactory;
-    protected DataSetClientServices dataSetClientServices;
-    protected DisplayerLocator displayerLocator;
-    protected DisplayerCoordinator displayerCoordinator;
-    protected PlaceManager placeManager;
     protected Event<TaskSelectionEvent> taskSelectionEvent;
     protected Event<TaskDashboardFocusEvent> taskDashboardFocusEvent;
-    protected DashboardI18n i18n;
 
     protected MetricDisplayer totalMetric;
     protected MetricDisplayer createdMetric;
@@ -125,8 +98,6 @@ public class TaskDashboard implements IsWidget {
     protected AbstractDisplayer tasksByStatus;
     protected TableDisplayer tasksTable;
 
-    protected MetricDisplayer selectedMetric = null;
-    protected String selectedProcess = null;
     protected List<Displayer> metricsGroup = new ArrayList<Displayer>();
     protected List<Displayer> metricsGroupOptional = new ArrayList<Displayer>();
     protected List<Displayer> chartsGroup = new ArrayList<Displayer>();
@@ -142,15 +113,9 @@ public class TaskDashboard implements IsWidget {
                          final PlaceManager placeManager,
                          final Event<TaskSelectionEvent> taskSelectionEvent,
                          final Event<TaskDashboardFocusEvent> taskDashboardFocusEvent) {
+        super(dashboardFactory, dataSetClientServices, placeManager, view.getI18nService(), processBreadCrumb, displayerLocator, displayerCoordinator);
 
         this.view = view;
-        this.i18n = view.getI18nService();
-        this.processBreadCrumb = processBreadCrumb;
-        this.dashboardFactory = dashboardFactory;
-        this.dataSetClientServices = dataSetClientServices;
-        this.displayerLocator = displayerLocator;
-        this.displayerCoordinator = displayerCoordinator;
-        this.placeManager = placeManager;
         this.taskSelectionEvent = taskSelectionEvent;
         this.taskDashboardFocusEvent = taskDashboardFocusEvent;
 
@@ -188,7 +153,7 @@ public class TaskDashboard implements IsWidget {
                 tasksByEndDate = createDisplayer(DashboardKpis.tasksByEndDate(i18n)),
                 tasksByRunningTime = createDisplayer(DashboardKpis.tasksByRunningTime(i18n)),
                 tasksByStatus = createDisplayer(DashboardKpis.tasksByStatus(i18n)),
-                tasksTable = createTableDisplayer(DashboardKpis.tasksTable(i18n)));
+                tasksTable = createTableDisplayer(DashboardKpis.tasksTable(i18n), COLUMN_TASK_DURATION, new DurationFormatter(COLUMN_TASK_CREATED_DATE, COLUMN_TASK_END_DATE)));
 
         metricsGroup.add(totalMetric);
         metricsGroup.add(readyMetric);
@@ -231,14 +196,6 @@ public class TaskDashboard implements IsWidget {
                     }
                 }
         );
-    }
-
-    public Widget asWidget() {
-        return view.asWidget();
-    }
-
-    public ProcessBreadCrumb getProcessBreadCrumb() {
-        return processBreadCrumb;
     }
 
     public MetricDisplayer getTotalMetric() {
@@ -313,74 +270,9 @@ public class TaskDashboard implements IsWidget {
         return tasksTable;
     }
 
-    public MetricDisplayer getSelectedMetric() {
-        return selectedMetric;
-    }
-
-    public String getSelectedProcess() {
-        return selectedProcess;
-    }
-
-    public MetricDisplayer createMetricDisplayer(DisplayerSettings settings) {
-        checkNotNull("displayerSettings", settings);
-        MetricDisplayer metricDisplayer = dashboardFactory.createMetricDisplayer();
-        metricDisplayer.setDisplayerSettings(settings);
-        metricDisplayer.setDataSetHandler(new DataSetHandlerImpl(dataSetClientServices, settings.getDataSetLookup()));
-        return metricDisplayer;
-    }
-
-    public AbstractDisplayer createDisplayer(DisplayerSettings settings) {
-        checkNotNull("displayerSettings", settings);
-        return (AbstractDisplayer) displayerLocator.lookupDisplayer(settings);
-    }
-
-    public TableDisplayer createTableDisplayer(DisplayerSettings settings) {
-        checkNotNull("displayerSettings", settings);
-        final TableDisplayer tableDisplayer = dashboardFactory.createTableDisplayer();
-        tableDisplayer.setDisplayerSettings(settings);
-        tableDisplayer.setDataSetHandler(new DataSetHandlerImpl(dataSetClientServices, settings.getDataSetLookup()));
-        tableDisplayer.addFormatter(COLUMN_TASK_DURATION, new DurationFormatter(COLUMN_TASK_CREATED_DATE, COLUMN_TASK_END_DATE));
-        tableDisplayer.setOnCellSelectedCommand(new Command() {
-            public void execute() {
-                tableCellSelected(tableDisplayer.getSelectedCellColumn(), tableDisplayer.getSelectedCellRow());
-            }
-        });
-        return tableDisplayer;
-    }
-
-    public void resetCurrentMetric() {
-        selectedMetric = null;
-        updateHeaderText();
-    }
-
-    public void changeCurrentMetric(MetricDisplayer metric) {
-        if (metric.isFilterOn()) {
-
-            // Reset existing metric selected as only a single metric can be filtered at the same time
-            if (selectedMetric != null && selectedMetric != metric) {
-                selectedMetric.filterReset();
-            }
-            // Set the selected metric as active
-            selectedMetric = metric;
-
-            // Update the header text
-            updateHeaderText();
-        } else {
-            selectedMetric = null;
-            updateHeaderText();
-        }
-    }
-
-    public void changeCurrentProcess(String name) {
-        selectedProcess = name;
-        updateHeaderText();
-        view.showBreadCrumb(name);
-    }
-
-    public void resetCurrentProcess() {
-        selectedProcess = null;
-        updateHeaderText();
-        view.hideBreadCrumb();
+    @Override
+    public AbstractDashboard.View getView() {
+        return view;
     }
 
     public void resetProcessBreadcrumb() {
@@ -407,35 +299,26 @@ public class TaskDashboard implements IsWidget {
             String status = "";
             if (createdMetric == selectedMetric) {
                 status = i18n.taskStatusCreated();
-            }
-            else if (readyMetric == selectedMetric) {
+            } else if (readyMetric == selectedMetric) {
                 status = i18n.taskStatusReady();
-            }
-            else if (reservedMetric == selectedMetric) {
+            } else if (reservedMetric == selectedMetric) {
                 status = i18n.taskStatusReserved();
-            }
-            else if (inProgressMetric == selectedMetric) {
+            } else if (inProgressMetric == selectedMetric) {
                 status = i18n.taskStatusInProgress();
-            }
-            else if (suspendedMetric == selectedMetric) {
+            } else if (suspendedMetric == selectedMetric) {
                 status = i18n.taskStatusSuspended();
-            }
-            else if (completedMetric == selectedMetric) {
+            } else if (completedMetric == selectedMetric) {
                 status = i18n.taskStatusCompleted();
-            }
-            else if (failedMetric == selectedMetric) {
+            } else if (failedMetric == selectedMetric) {
                 status = i18n.taskStatusFailed();
-            }
-            else if (errorMetric == selectedMetric) {
+            } else if (errorMetric == selectedMetric) {
                 status = i18n.taskStatusError();
-            }
-            else if (exitedMetric == selectedMetric) {
+            } else if (exitedMetric == selectedMetric) {
                 status = i18n.taskStatusExited();
-            }
-            else if (obsoleteMetric == selectedMetric) {
+            } else if (obsoleteMetric == selectedMetric) {
                 status = i18n.taskStatusObsolete();
             }
-            view.setHeaderText(i18n.selectedTaskStatusHeader(status , selectedProcess));
+            view.setHeaderText(i18n.selectedTaskStatusHeader(status, selectedProcess));
         }
     }
